Renames bchydro_epistemic.py to bchydro_2016_epistemic.py
