+++ openquake/hazardlib/tests/gsim/bchydro_2016_epistemic_test.py
@@ -16,12 +16,9 @@
 # You should have received a copy of the GNU Affero General Public License
 # along with OpenQuake. If not, see <http://www.gnu.org/licenses/>
 
-from openquake.hazardlib.gsim.bchydro_epistemic import (BCHydroSERASInter,
-                                                        BCHydroSERASInterHigh,
-                                                        BCHydroSERASInterLow,
-                                                        BCHydroSERASSlab,
-                                                        BCHydroSERASSlabHigh,
-                                                        BCHydroSERASSlabLow)
+from openquake.hazardlib.gsim.bchydro_2016_epistemic import (
+    BCHydroSERASInter, BCHydroSERASInterHigh, BCHydroSERASInterLow,
+    BCHydroSERASSlab, BCHydroSERASSlabHigh, BCHydroSERASSlabLow)
 from openquake.hazardlib.tests.gsim.utils import BaseGSIMTestCase
